Move Token & TokenType to separated files.
