+++ src/token.cpp
@@ -12,24 +12,10 @@
 	#include <utility>
 #endif
 
-/* ------------------------- Относится к TokenType -------------------------- */
-void token::swap(token::TokenType &lhs, token::TokenType &rhs)
-{
-	std::swap(rhs.type, lhs.type);
-	std::swap(rhs.precedence, lhs.precedence);
-	std::swap(rhs.association, lhs.association);
-}
+// Объявление функции, из файла tokentype.hpp
+void token::swap(token::TokenType &lhs, token::TokenType &rhs);
 
-bool token::operator==(const token::TokenType &rhs, const token::TokenType &lhs)
-{
-	return static_cast<int8_t>(rhs.type) == static_cast<int8_t>(lhs.type)
-		   && static_cast<unsigned int>(rhs.precedence)
-				  == static_cast<unsigned int>(lhs.precedence)
-		   && static_cast<int8_t>(rhs.association)
-				  == static_cast<int8_t>(lhs.association);
-}
-
-/* --------------------------- Относится к Token ---------------------------- */
+/* -------------------------------------------------------------------------- */
 bool token::operator==(const Token &rhs, const Token &lhs)
 {
 	return rhs.m_value == lhs.m_value && rhs.m_type == lhs.m_type;
